+++ e2e/nx-verdaccio-e2e/test/target-env-setup.e2e.test.ts
@@ -1,20 +1,10 @@
-import type { Tree } from '@nx/devkit';
-import { join } from 'node:path';
 import { afterEach, expect } from 'vitest';
-import {
-  addJsLibToWorkspace,
-  materializeTree,
-  registerPluginInWorkspace,
-} from '@push-based/test-nx-utils';
-import { updateProjectConfiguration } from 'nx/src/generators/utils/project-configuration';
-import { executeProcess, teardownTestFolder } from '@push-based/test-utils';
+import { executeProcess } from '@push-based/test-utils';
 import { TARGET_ENVIRONMENT_SETUP } from '@push-based/nx-verdaccio';
 
 describe('nx-verdaccio plugin nxv-env-setup target', () => {
-  let tree: Tree;
   const projectA = 'lib-a';
   const projectAE2e = `${projectA}-e2e`;
-  const e2eProjectARoot = join('projects', projectAE2e);
   const baseDir = `tmp/environments/${process.env['NX_TASK_TARGET_PROJECT']}/__test__/target-env-setup`;
 
   afterEach(async () => {
