+++ projects/nx-verdaccio/src/executors/env-setup/executor.ts
@@ -20,7 +20,8 @@ import {
 import { runSingleExecutor } from '../../internal/run-executor';
 import { getEnvironmentRoot } from '../../internal/environment-root';
 import { cleanupEnv } from '../internal/cleanup-env';
-import { setupNpmWorkspace } from '../env-bootstrap/npm';
+
+import { setupNpmWorkspace } from './npm';
 
 export type ExecutorOutput = {
   success: boolean;
